Adding in Reflection-based testing, to ensure that all properties are covered.

+++ OpenSim/Data/Tests/BasicGridTest.cs
@@ -49,9 +49,13 @@ namespace OpenSim.Data.Tests
         public void removeAllRegions()
         {
             // Clean up all the regions.
-            foreach (RegionProfileData region in db.GetRegionsByName("", 100))
+            List<RegionProfileData> regions = db.GetRegionsByName("", 100);
+            if(regions != null)
             {
-                db.DeleteProfile(region.Uuid.ToString());
+                foreach (RegionProfileData region in regions)
+                {
+                    db.DeleteProfile(region.Uuid.ToString());
+                }
             }
         }
 
@@ -74,35 +78,9 @@ namespace OpenSim.Data.Tests
         protected RegionProfileData createRegion(UUID regionUUID, string regionName)
         {
             RegionProfileData reg = new RegionProfileData();
+            ScrambleForTesting.Scramble(reg);
             reg.Uuid = regionUUID;
             reg.RegionName = regionName;
-            reg.RegionHandle = (ulong) random.Next();
-            reg.RegionLocX = (uint) random.Next();
-            reg.RegionLocY = (uint) random.Next();
-            reg.RegionLocZ = (uint) random.Next();
-            reg.RegionSendKey = RandomName();
-            reg.RegionRecvKey = RandomName();
-            reg.RegionSecret = RandomName();
-            reg.RegionOnline = false;
-            reg.ServerIP = RandomName();
-            reg.ServerPort = (uint) random.Next();
-            reg.ServerURI = RandomName();
-            reg.ServerHttpPort = (uint) random.Next();
-            reg.ServerRemotingPort = (uint) random.Next();
-            reg.NorthOverrideHandle = (ulong) random.Next();
-            reg.SouthOverrideHandle = (ulong) random.Next();
-            reg.EastOverrideHandle = (ulong) random.Next();
-            reg.WestOverrideHandle = (ulong) random.Next();
-            reg.RegionDataURI = RandomName();
-            reg.RegionAssetURI = RandomName();
-            reg.RegionAssetSendKey = RandomName();
-            reg.RegionAssetRecvKey = RandomName();
-            reg.RegionUserURI = RandomName();
-            reg.RegionUserSendKey = RandomName();
-            reg.RegionUserRecvKey = RandomName();
-            reg.RegionMapTextureID = UUID.Random();
-            reg.Owner_uuid = UUID.Random();
-            reg.OriginUUID = UUID.Random();
 
             db.AddProfile(reg);
 
@@ -119,47 +97,12 @@ namespace OpenSim.Data.Tests
         }
 
         [Test]
-        public void T999_StillNull()
-        {
-            Assert.That(db.GetProfileByUUID(zero),Is.Null);
-        }
-
-        [Test]
         public void T011_AddRetrieveCompleteTest()
         {
             RegionProfileData newreg = createRegion(region2, "|<Goth@m Ci1y>|");
             RegionProfileData retreg = db.GetProfileByUUID(region2);
 
-            Assert.That(retreg.RegionName, Is.EqualTo(newreg.RegionName), "Assert.That(retreg.RegionName, Is.EqualTo(newreg.RegionName))");
-            Assert.That(retreg.Uuid, Is.EqualTo(region2), "Assert.That(retreg.Uuid, Is.EqualTo(region2))");
-            Assert.That(retreg.RegionHandle, Is.EqualTo(newreg.RegionHandle), "Assert.That(retreg.RegionHandle, Is.EqualTo(newreg.RegionHandle))");
-            Assert.That(retreg.RegionLocX, Is.EqualTo(newreg.RegionLocX), "Assert.That(retreg.RegionLocX, Is.EqualTo(newreg.RegionLocX))");
-            Assert.That(retreg.RegionLocY, Is.EqualTo(newreg.RegionLocY), "Assert.That(retreg.RegionLocY, Is.EqualTo(newreg.RegionLocY))");
-            Assert.That(retreg.RegionLocZ, Is.EqualTo(newreg.RegionLocZ), "Assert.That(retreg.RegionLocZ, Is.EqualTo(newreg.RegionLocZ))");
-            Assert.That(retreg.RegionSendKey, Is.EqualTo(newreg.RegionSendKey), "Assert.That(retreg.RegionSendKey, Is.EqualTo(newreg.RegionSendKey))");
-            Assert.That(retreg.RegionRecvKey, Is.EqualTo(newreg.RegionRecvKey), "Assert.That(retreg.RegionRecvKey, Is.EqualTo(newreg.RegionRecvKey))");
-            Assert.That(retreg.RegionSecret, Is.EqualTo(newreg.RegionSecret), "Assert.That(retreg.RegionSecret, Is.EqualTo(newreg.RegionSecret))");
-            Assert.That(retreg.RegionOnline, Is.EqualTo(newreg.RegionOnline), "Assert.That(retreg.RegionOnline, Is.EqualTo(newreg.RegionOnline))");
-            Assert.That(retreg.OriginUUID, Is.EqualTo(newreg.OriginUUID), "Assert.That(retreg.OriginUUID, Is.EqualTo(newreg.OriginUUID))");
-            Assert.That(retreg.ServerIP, Is.EqualTo(newreg.ServerIP), "Assert.That(retreg.ServerIP, Is.EqualTo(newreg.ServerIP))");
-            Assert.That(retreg.ServerPort, Is.EqualTo(newreg.ServerPort), "Assert.That(retreg.ServerPort, Is.EqualTo(newreg.ServerPort))");
-            Assert.That(retreg.ServerURI, Is.EqualTo(newreg.ServerURI), "Assert.That(retreg.ServerURI, Is.EqualTo(newreg.ServerURI))");
-            Assert.That(retreg.ServerHttpPort, Is.EqualTo(newreg.ServerHttpPort), "Assert.That(retreg.ServerHttpPort, Is.EqualTo(newreg.ServerHttpPort))");
-            Assert.That(retreg.ServerRemotingPort, Is.EqualTo(newreg.ServerRemotingPort), "Assert.That(retreg.ServerRemotingPort, Is.EqualTo(newreg.ServerRemotingPort))");
-            Assert.That(retreg.NorthOverrideHandle, Is.EqualTo(newreg.NorthOverrideHandle), "Assert.That(retreg.NorthOverrideHandle, Is.EqualTo(newreg.NorthOverrideHandle))");
-            Assert.That(retreg.SouthOverrideHandle, Is.EqualTo(newreg.SouthOverrideHandle), "Assert.That(retreg.SouthOverrideHandle, Is.EqualTo(newreg.SouthOverrideHandle))");
-            Assert.That(retreg.EastOverrideHandle, Is.EqualTo(newreg.EastOverrideHandle), "Assert.That(retreg.EastOverrideHandle, Is.EqualTo(newreg.EastOverrideHandle))");
-            Assert.That(retreg.WestOverrideHandle, Is.EqualTo(newreg.WestOverrideHandle), "Assert.That(retreg.WestOverrideHandle, Is.EqualTo(newreg.WestOverrideHandle))");
-            Assert.That(retreg.RegionDataURI, Is.EqualTo(newreg.RegionDataURI), "Assert.That(retreg.RegionDataURI, Is.EqualTo(newreg.RegionDataURI))");
-            Assert.That(retreg.RegionAssetURI, Is.EqualTo(newreg.RegionAssetURI), "Assert.That(retreg.RegionAssetURI, Is.EqualTo(newreg.RegionAssetURI))");
-            Assert.That(retreg.RegionAssetSendKey, Is.EqualTo(newreg.RegionAssetSendKey), "Assert.That(retreg.RegionAssetSendKey, Is.EqualTo(newreg.RegionAssetSendKey))");
-            Assert.That(retreg.RegionAssetRecvKey, Is.EqualTo(newreg.RegionAssetRecvKey), "Assert.That(retreg.RegionAssetRecvKey, Is.EqualTo(newreg.RegionAssetRecvKey))");
-            Assert.That(retreg.RegionUserURI, Is.EqualTo(newreg.RegionUserURI), "Assert.That(retreg.RegionUserURI, Is.EqualTo(newreg.RegionUserURI))");
-            Assert.That(retreg.RegionUserSendKey, Is.EqualTo(newreg.RegionUserSendKey), "Assert.That(retreg.RegionUserSendKey, Is.EqualTo(newreg.RegionUserSendKey))");
-            Assert.That(retreg.RegionUserRecvKey, Is.EqualTo(newreg.RegionUserRecvKey), "Assert.That(retreg.RegionUserRecvKey, Is.EqualTo(newreg.RegionUserRecvKey))");
-            Assert.That(retreg.RegionMapTextureID, Is.EqualTo(newreg.RegionMapTextureID), "Assert.That(retreg.RegionMapTextureID, Is.EqualTo(newreg.RegionMapTextureID))");
-            Assert.That(retreg.Owner_uuid, Is.EqualTo(newreg.Owner_uuid), "Assert.That(retreg.Owner_uuid, Is.EqualTo(newreg.Owner_uuid))");
-            Assert.That(retreg.OriginUUID, Is.EqualTo(newreg.OriginUUID), "Assert.That(retreg.OriginUUID, Is.EqualTo(newreg.OriginUUID))");
+            Assert.That(retreg, Constraints.PropertyCompareConstraint(newreg).IgnoreProperty(x => x.RegionOnline));
 
             retreg = db.GetProfileByHandle(newreg.RegionHandle);
             Assert.That(retreg.Uuid, Is.EqualTo(region2), "Assert.That(retreg.Uuid, Is.EqualTo(region2))");
@@ -218,6 +161,12 @@ namespace OpenSim.Data.Tests
             Assert.That(listreg[0].Uuid,Is.Not.EqualTo(listreg[1].Uuid), "Assert.That(listreg[0].Uuid,Is.Not.EqualTo(listreg[1].Uuid))");
             Assert.That(listreg[0].Uuid, Is.EqualTo(region1) | Is.EqualTo(region2), "Assert.That(listreg[0].Uuid, Is.EqualTo(region1) | Is.EqualTo(region2))");
             Assert.That(listreg[1].Uuid, Is.EqualTo(region1) | Is.EqualTo(region2), "Assert.That(listreg[1].Uuid, Is.EqualTo(region1) | Is.EqualTo(region2))");
+        }
+
+        [Test]
+        public void T999_StillNull()
+        {
+            Assert.That(db.GetProfileByUUID(zero), Is.Null);
         }
 
         protected static string RandomName()
